Update DOCUMENTATION.md
Updated documentation to include Pseudo-code and a techincal explanation of how the code is working.
Included examples of how the fractal structure can be generated with images and the required variables and a seed.
Included some of the challenges and how they were solved during the assignement workflow.
Updated the list of references to include information used in the solution of the assignment.